Remove syscall-overrides work
As I'm not using pread_fn, it no longer makes sense as part of this PR
and would be part of the finishing of 604 once that is unblocked.
