+++ src/enclave/enclave_mem.c
@@ -4,6 +4,8 @@
 
 #define _GNU_SOURCE
 
+#include <lkl.h>
+
 #include <sys/mman.h>
 #include <sys/types.h>
 #include <unistd.h>
@@ -28,6 +30,7 @@ static size_t mmap_num_pages;   // Total number of pages that can be mmap'ed
 
 static size_t used_pages =
     0; // Tracks the number of used pages for the mmap tracing
+static int mmap_files;
 
 #if DEBUG
 extern int sgxlkl_trace_mmap;
@@ -92,6 +95,61 @@ static void* index_to_addr(size_t index)
 static size_t addr_to_index(void* addr)
 {
     return ((char*)mmap_end - (char*)addr) / PAGE_SIZE;
+}
+
+long syscall_SYS_mmap(
+    void* addr,
+    size_t length,
+    int prot,
+    int flags,
+    int fd,
+    off_t offset)
+{
+    if ((flags & MAP_SHARED) && (flags & MAP_PRIVATE))
+    {
+        sgxlkl_warn("mmap() with MAP_SHARED and MAP_PRIVATE not supported\n");
+        return -EINVAL;
+    }
+    // Anonymous mapping/allocation
+    else if (fd == -1 && (flags & MAP_ANONYMOUS))
+    {
+        return (long)enclave_mmap(addr, length, flags & MAP_FIXED, prot, 1);
+    }
+    // File-backed mapping (if allowed)
+    else if ((fd >= 0) && enclave_mmap_files_flags_supported(flags))
+    {
+        void* mem =
+            enclave_mmap(addr, length, flags & MAP_FIXED, prot | PROT_WRITE, 0);
+
+        if (mem > 0)
+        {
+            // Read file into memory
+            size_t readb = 0;
+            ssize_t ret = 0;
+            while ((ret = lkl_sys_pread64(
+                        fd, ((char*)mem) + readb, length - readb, offset)) > 0)
+            {
+                readb += ret;
+                offset += ret;
+            };
+
+            if (ret < 0)
+            {
+                enclave_munmap(addr, length);
+                return -EBADF;
+            }
+
+            // Set requested page permissions
+            if ((prot | PROT_WRITE) != prot)
+                mprotect(mem, length, prot);
+        }
+
+        return (long)mem;
+    }
+    else
+    {
+        return -EINVAL;
+    }
 }
 
 void* syscall_SYS_mremap(
@@ -178,6 +236,22 @@ void enclave_mman_init(const void* base, size_t num_pages, int _mmap_files)
     bitmap_set(mmap_fresh_bitmap, 0, mmap_num_pages);
 
     mmap_files = _mmap_files;
+}
+
+/*
+ * Returns 1 if we can mmap files using the given flags
+ * returns 0 otherwise.
+ */
+int enclave_mmap_files_flags_supported(int flags)
+{
+    int supported_flags = 0;
+
+    if (mmap_files == ENCLAVE_MMAP_FILES_SHARED)
+        supported_flags = MAP_PRIVATE | MAP_SHARED;
+    else if (mmap_files == ENCLAVE_MMAP_FILES_PRIVATE)
+        supported_flags = MAP_PRIVATE;
+
+    return supported_flags & flags;
 }
 
 /*
